Refactor and update OAuth2-related classes and handlers
