--- backend/src/main/java/SniffStep/common/Failure.java
@@ -1,10 +1,0 @@
-package SniffStep.common;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class Failure implements Result {
-    private String msg;
-}
