--- backend/src/main/java/SniffStep/common/Result.java
@@ -1,4 +1,0 @@
-package SniffStep.common;
-
-public interface Result {
-}
