--- backend/src/main/java/SniffStep/common/Success.java
@@ -1,12 +1,0 @@
-package SniffStep.common;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@Getter
-@AllArgsConstructor
-public class Success<T> implements Result {
-    private T data;
-}
