+++ backend/src/main/java/SniffStep/common/config/oauth/handler/OAuthLoginFailureHandler.java
@@ -12,7 +12,7 @@ import java.io.IOException;
 
 @Slf4j
 @Component
-public class OAuth2LoginFailureHandler implements AuthenticationFailureHandler {
+public class OAuthLoginFailureHandler implements AuthenticationFailureHandler {
 
     @Override
     public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response, AuthenticationException exception) throws IOException, ServletException {
