+++ backend/src/main/java/SniffStep/controller/AuthController.java
@@ -1,17 +1,14 @@
 package SniffStep.controller;
 
-import SniffStep.common.Response;
-import SniffStep.common.jwt.service.JwtService;
 import SniffStep.dto.auth.LoginDTO;
 import SniffStep.dto.auth.SignUpRequestDTO;
 import SniffStep.service.AuthService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.*;
-
-import static SniffStep.common.Response.success;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 
 @RequiredArgsConstructor
@@ -20,19 +17,18 @@ import static SniffStep.common.Response.success;
 public class AuthController {
 
     private final AuthService authService;
-    private final JwtService jwtService;
 
-    @ResponseStatus(HttpStatus.CREATED)
+
+    // 자체 회원가입
     @PostMapping("/signup")
-    public Response signup(@RequestBody SignUpRequestDTO signUpRequestDTO) throws Exception {
+    public ResponseEntity signup(@RequestBody SignUpRequestDTO signUpRequestDTO) throws Exception {
         authService.signup(signUpRequestDTO);
-        return success();
+        return ResponseEntity.ok().build();
     }
 
+    // 로그인
     @PostMapping("/login")
-    public ResponseEntity<Response> login(@RequestBody LoginDTO loginDTO) {
-        return ResponseEntity.ok(success(authService.login(loginDTO)));
+    public ResponseEntity login(@RequestBody LoginDTO loginDTO) {
+        return ResponseEntity.ok(authService.login(loginDTO));
     }
-
-
 }
