+++ backend/src/main/java/SniffStep/controller/OAuthController.java
@@ -1,15 +1,11 @@
 package SniffStep.controller;
 
 import SniffStep.common.config.oauth.GetSocialOAuthRes;
-import SniffStep.common.config.oauth.OAuthService;
+import SniffStep.service.OAuthService;
 import SniffStep.common.jwt.service.JwtService;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.io.IOException;
