+++ backend/src/main/java/SniffStep/service/OAuthService.java
@@ -1,6 +1,7 @@
-package SniffStep.common.config.oauth;
+package SniffStep.service;
 
 
+import SniffStep.common.config.oauth.*;
 import SniffStep.common.jwt.dto.TokenDto;
 import SniffStep.common.jwt.service.JwtService;
 import SniffStep.entity.JwtTokenType;
